% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{plotImpliedVol}
\alias{plotImpliedVol}
\title{A function that plots the implied volatility from a data frame
given}
\usage{
plotImpliedVol(df, r = 0.03, labelPoints = FALSE)
}
\arguments{
\item{df}{The data frame containing information of your options. It
must follow the format df = data.frame( strike = c(50, 20),
type = c("C", "P"), optionPrice = c(1.62, 0.01), futurePrice =
c(48.03, 48.03), time_to_expiry = c(0.1423, 0.1423) )}

\item{r}{= 0.03 The default risk free interest rate is set to 3
percents. You may alter it if necessary}

\item{labelPoints}{= FALSE By default, point labels are not
included. If set to TRUE, a string in the order of
"strike, type, optionPrice, futurePrice, time_to_expiry" will
be attached to each point.}
}
\value{
vols: A vector containing implied votalities, in the order
    of input data frame, df
}
\description{
A function that plots the implied volatility from a data frame
given
}
