% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{putPrice}
\alias{putPrice}
\title{A function that updates put price accoridng to black 76 formula}
\usage{
putPrice(Vol, F, K, T, r)
}
\arguments{
\item{Vol}{Volatility, in the context of this package, it is a trial value}

\item{F}{Future price of the option}

\item{K}{Strike price of the option}

\item{T}{Time to expiry in year}

\item{r}{Risk free interst rate}
}
\value{
p: The put price
}
\description{
A function that updates put price accoridng to black 76 formula
}
